knitter: wait for both selectors to clear the limit
We now maintain a "current line direction" that allows us
to know whether the carriage is on the expected side of the
working area for the line to be considered complete.

Then we get rid of m_workedOnLine that is no longer useful.

We also use the current line direction when determining which
needle to program, making this more reliable than depending on
the instantaneous carriage direction.

Had to fix a few tests that used a mix of Knit and Garter carriage.