encoders: fix carriage detection on the right
Mostly by copying the logic from encA_rising into encA_falling, and
special-casing the KH910 with its inverted K-only reading.

Also, if we have detected a Garter carriage, we already decided that it's
too hard to try to reset the position when it crosses the turn marks.
If we punt on resetting the belt shift as well, we save some logic and
fix the few weird spots where it turns while its outer magnet is right
on the turn mark.