Track position as a 16-bit signed integer
This avoids overflows on the left and right.

Because 16-bit accesses are not atomic, the update of
`Knitter` members `m_position`, `m_direction` from `Encoders`
is no longer done in the ISR  but in a `copyEncodersState` method
that is called regularly from methods dispatched by the FSM. This had
quite a bit of impact on the tests unfortunately.

A bounds test was added to avoid accessing (and sendind to solenoids!)
out-of-bounds pixels.